refactor(console): extract create deployment confirm in hook
