+++ apps/deploy-web/src/components/new-deployment/ManifestEdit.tsx
@@ -39,11 +39,8 @@ import { LinkTo } from "../shared/LinkTo";
 import { PrerequisiteList } from "../shared/PrerequisiteList";
 import ViewPanel from "../shared/ViewPanel";
 import { SdlBuilder, SdlBuilderRefType } from "./SdlBuilder";
-import { usePopup } from "@akashnetwork/ui/context";
-import { LeaseSpecDetail } from "../shared/LeaseSpecDetail";
-import { v3Manifest, v2Service, v2ManifestService, v3ManifestService } from "@akashnetwork/akashjs/build/sdl/types";
-import first from "lodash/first";
 import { importSimpleSdl } from "@src/utils/sdl/sdlImport";
+import { useManagedDeploymentConfirm } from "@src/hooks/useManagedDeploymentConfirm";
 
 type Props = {
   onTemplateSelected: Dispatch<TemplateCreation | null>;
@@ -77,7 +74,7 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
   const fileUploadRef = useRef<HTMLInputElement>(null);
   const wallet = useWallet();
   const managedDenom = useManagedWalletDenom();
-  const { confirm } = usePopup();
+  const { createDeploymentConfirm } = useManagedDeploymentConfirm();
 
   useWhen(wallet.isManaged && sdlDenom === "uakt", () => {
     setSdlDenom(managedDenom);
@@ -190,31 +187,12 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
     if (isManaged) {
       const services = importSimpleSdl(editedManifest as string);
 
-      if (!deploymentData) {
+      if (!services) {
         setParsingError("Error while parsing SDL file");
         return;
       }
 
-      const isConfirmed = await confirm({
-        title: "Confirm deployment creation?",
-        message: (
-          <div className="space-y-2">
-            {services.map(service => {
-              return (
-                <div key={service.image} className="rounded border p-4">
-                  <div className="mb-2 text-sm"><span className="font-bold">{service.title}</span>:{service.image}</div>
-                  <div className="flex items-center space-x-4 whitespace-nowrap">
-                    <LeaseSpecDetail type="cpu" className="flex-shrink-0" value={service.profile?.cpu as number} />
-                    {service.profile?.hasGpu && <LeaseSpecDetail type="gpu" className="flex-shrink-0" value={service.profile?.gpu as number} />}
-                    <LeaseSpecDetail type="ram" className="flex-shrink-0" value={`${service.profile?.ram} ${service.profile?.ramUnit}`} />
-                    <LeaseSpecDetail type="storage" className="flex-shrink-0" value={`${service.profile?.storage} ${service.profile?.storageUnit}`} />
-                  </div>
-                </div>
-              );
-            })}
-          </div>
-        )
-      });
+      const isConfirmed = await createDeploymentConfirm(services);
 
       if (isConfirmed) {
         await handleCreateClick(defaultDeposit, envConfig.NEXT_PUBLIC_MASTER_WALLET_ADDRESS);
